Add rc.local_pitrex, to copy wpa_supplicant from /boot/settings
Rather than add a whole bunch to rc.local, and need to deal with the
'exit 0' at the end, source in our own bootup. I'm not wedded to this
approach, we can change it later if we decide to.

The core change introduced here is that wpa_supplicant is generated by
packer into the /boot/settings/ directory, so a power user can edit it
there from whatever OS they are using. A cheap way of having
a downloadable image that doesn't need booting into Linux to change the
wifi details.

I've also made the 'performance' and 'hdmi' setting modes run-time
optional now, via flag files in /boot/settings.